+++ node/koa-router.js
@@ -1,7 +1,7 @@
 import { Readable } from "node:stream";
 
 import { FetchRouter } from "../core/fetch-router.js";
-import { getFetchRequest } from "./node-router.js";
+import { getFetchRequest, getResponseReadable } from "./node-router.js";
 
 /** @typedef {import("koa").Context} Context */
 /** @typedef {import("./node-router.js").NodeRouterOptions} NodeRouterOptions */
@@ -54,7 +54,7 @@ export class KoaRouter {
 		}
 
 		if (response.body) {
-			ctx.response.body = Readable.fromWeb(response.body);
+			ctx.response.body = getResponseReadable(response);
 		}
 	}
 }